Fix version and refactor
The version of the crud has been fixed, since the example did not work
due to the fact that apparently an early version of the 0.6.0 crud was
used, since the version was used before the release of this version took
place. Also added a description of the limitations of the crud and
cartridge-cli versions. Fixed style code that forced `spring-javaformat:
apply`.

Closes #2